146384:samba-4.2.3-alt1 +3 more
- 4.2.3
- add alternatives for libwbclient

libtevent-0.9.25-alt1

- 0.9.25

tdb-1.3.6-alt1

- 1.3.6

sssd-1.13.0-alt1

- 1.13.0
- add alternatives for libwbclient
- add alternatives for cifs-idmap plugin
- use _sssd user for run services

+++ s/samba/samba4.spec
@@ -1,4 +1,5 @@
 %define _localstatedir /var
+%define libwbc_alternatives_version 0.12
 
 # internal libs
 %def_without talloc
@@ -32,7 +33,7 @@
 %def_with libcephfs
 
 Name: samba
-Version: 4.2.0
+Version: 4.2.3
 Release: alt1
 Group: System/Servers
 Summary: The Samba4 CIFS and AD client and server suite
@@ -70,6 +71,7 @@ Obsoletes: samba4-doc < %version-%release
 
 Requires(pre): %name-common = %version-%release
 Requires: %name-libs = %version-%release
+Requires: %name-common-tools = %version-%release
 %if_with libwbclient
 Requires: libwbclient = %version-%release
 %endif
@@ -77,6 +79,7 @@ Requires: libwbclient = %version-%release
 BuildRequires: libe2fs-devel
 BuildRequires: libacl-devel
 BuildRequires: libaio-devel
+BuildRequires: libarchive-devel
 BuildRequires: libattr-devel
 BuildRequires: libncurses-devel
 BuildRequires: libpam-devel
@@ -95,17 +98,18 @@ BuildRequires: libiniparser-devel
 BuildRequires: libkrb5-devel libssl-devel libcups-devel
 BuildRequires: gawk libgtk+2-devel libcap-devel libuuid-devel
 BuildRequires: inkscape libxslt xsltproc netpbm dblatex html2text docbook-style-xsl
-%{?_without_talloc:BuildRequires: libtalloc-devel >= 2.0.8 libpytalloc-devel}
-%{?_without_tevent:BuildRequires: libtevent-devel >= 0.9.18 python-module-tevent}
-%{?_without_tdb:BuildRequires: libtdb-devel >= 1.2.11  python-module-tdb}
-%{?_without_ntdb:BuildRequires: libntdb-devel >= 0.9  python-module-ntdb}
-%{?_without_ldb:BuildRequires: libldb-devel >= 1.1.14 python-module-pyldb-devel}
-%{?_with_clustering_support:BuildRequires: ctdb-devel}
+%{?_without_talloc:BuildRequires: libtalloc-devel >= 2.1.2 libpytalloc-devel}
+%{?_without_tevent:BuildRequires: libtevent-devel >= 0.9.25 python-module-tevent}
+%{?_without_tdb:BuildRequires: libtdb-devel >= 1.3.6  python-module-tdb}
+%{?_without_ntdb:BuildRequires: libntdb-devel >= 1.0  python-module-ntdb}
+%{?_without_ldb:BuildRequires: libldb-devel >= 1.1.20 python-module-pyldb-devel}
+#{?_with_clustering_support:BuildRequires: ctdb-devel}
 %{?_with_testsuite:BuildRequires: ldb-tools}
 %{?_with_systemd:BuildRequires: libsystemd-devel}
 %{?_enable_avahi:BuildRequires: libavahi-devel}
 %{?_enable_glusterfs:BuildRequires: glusterfs3-devel >= 3.4.0.16}
 %{?_with_libcephfs:BuildRequires: ceph-devel}
+%{?_with_dc:BuildRequires: libgnutls-devel}
 BuildRequires: perl-Perl4-CoreLibs
 
 %description
@@ -114,7 +118,8 @@ Samba is the standard Windows interoperability suite of programs for Linux and U
 %package client
 Summary: Samba client programs
 Group: Networking/Other
-Requires: %name-common = %version-%release
+Requires(pre): %name-common = %version-%release
+Requires: %name-common-tools = %version-%release
 Requires: %name-client-libs = %version-%release
 %if_with libsmbclient
 Requires: libsmbclient = %version-%release
@@ -133,6 +138,7 @@ of SMB/CIFS shares and printing to SMB/CIFS printers.
 Summary: Samba client libraries
 Group: Networking/Other
 Conflicts: samba-common < %version-%release
+Requires(pre): %name-common = %version-%release
 
 %description client-libs
 The samba-client-libs package contains internal libraries needed by the
@@ -153,7 +159,7 @@ packages of Samba.
 %package common-libs
 Summary: Libraries used by both Samba servers and clients
 Group: System/Libraries
-Requires: %name-common = %version-%release
+Requires(pre): %name-common = %version-%release
 Requires: %name-client-libs = %version-%release
 %if_with libwbclient
 Requires: libwbclient = %version-%release
@@ -247,7 +253,7 @@ Summary: The SMB client library
 Group: System/Libraries
 Provides: libsmbclient4 = %version-%release
 Obsoletes: libsmbclient4 < %version-%release
-Requires: %name-common = %version-%release
+Requires(pre): %name-common = %version-%release
 Requires: %name-client-libs = %version-%release
 
 %description -n libsmbclient
@@ -355,7 +361,7 @@ samba4-test provides testing tools for both the server and client
 packages of Samba.
 
 %package test-libs
-Summary: Libraries need by teh testing tools for Samba servers and clients
+Summary: Libraries need by the testing tools for Samba servers and clients
 Group: System/Libraries
 Requires: %name-libs = %version-%release
 
@@ -378,7 +384,8 @@ packages of Samba.
 %package winbind
 Summary: Samba winbind
 Group: System/Servers
-Requires: %name-common = %version-%release
+Requires(pre): %name-common = %version-%release
+Requires: %name-common-tools = %version-%release
 Requires: %name-libs = %version-%release
 Provides: samba4-winbind = %version-%release
 Obsoletes: samba4-winbind < %version-%release
@@ -522,7 +529,7 @@ and use CTDB instead.
 
 %define _samba4_libraries heimdal,!zlib,!popt%{_talloc_lib}%{_tevent_lib}%{_tdb_lib}%{_ntdb_lib}%{_ldb_lib}
 
-%define _samba4_idmap_modules idmap_ad,idmap_rid,idmap_adex,idmap_hash,idmap_tdb2
+%define _samba4_idmap_modules idmap_ad,idmap_rid,idmap_adex,idmap_hash,idmap_tdb2,idmap_ldap
 %define _samba4_pdb_modules pdb_tdbsam,pdb_ldap,pdb_ads,pdb_smbpasswd,pdb_wbc_sam,pdb_samba4
 %define _samba4_auth_modules auth_unix,auth_wbc,auth_server,auth_netlogond,auth_script,auth_samba4
 # auth_domain needs to be static
@@ -530,7 +537,7 @@ and use CTDB instead.
 
 %define _libsmbclient %nil
 %if_without libsmbclient
-%define _libsmbclient smbclient,smbsharemodes,
+%define _libsmbclient smbclient,
 %endif
 
 %define _libwbclient %nil
@@ -588,7 +595,8 @@ and use CTDB instead.
 	--with-profiling-data \
 %endif
 	%{subst_enable avahi} \
-	--disable-gnutls \
+	%{subst_enable glusterfs} \
+	--disable-rpath \
 	--disable-rpath-install
 
 %make_build
@@ -611,7 +619,28 @@ mkdir -p %buildroot%_pkgconfigdir
 mkdir -p %buildroot%_initdir
 mkdir -p %buildroot%_unitdir
 mkdir -p %buildroot%_sysconfdir/{pam.d,logrotate.d,security,sysconfig}
-mkdir -p %buildroot/lib/tmpfiles.d
+mkdir -p %buildroot%_tmpfilesdir
+
+# Move libwbclient.so* into private directory, it cannot be just libdir/samba
+# because samba uses rpath with this directory.
+install -d -m 0755 %buildroot%_libdir/samba/wbclient
+mv %buildroot%_libdir/libwbclient.so* %buildroot%_libdir/samba/wbclient
+if [ ! -f %buildroot%_libdir/samba/wbclient/libwbclient.so.%libwbc_alternatives_version ]
+then
+    echo "Expected libwbclient version not found, please check if version has changed."
+    exit -1
+fi
+ln -s ../..%_libdir/samba/wbclient/libwbclient.so.%libwbc_alternatives_version %buildroot%_libdir/
+ln -s ../..%_libdir/samba/wbclient/libwbclient.so.0 %buildroot%_libdir/
+ln -s ../..%_libdir/samba/wbclient/libwbclient.so %buildroot%_libdir/
+
+# Add alternatives for libwbclient
+mkdir -p %buildroot%_altdir
+printf '%_libdir/libwbclient.so.%libwbc_alternatives_version\t%_libdir/samba/wbclient/libwbclient.so.%libwbc_alternatives_version\t10\n' > %buildroot%_altdir/libwbclient-samba
+printf '%_libdir/libwbclient.so.0\t%_libdir/samba/wbclient/libwbclient.so.0\t10\n' >> %buildroot%_altdir/libwbclient-samba
+
+printf '%_libdir/libwbclient.so\t%_libdir/samba/wbclient/libwbclient.so\t10\n' > %buildroot%_altdir/libwbclient-devel-samba
+
 
 # Install other stuff
 install -m644 %SOURCE1 %buildroot%_sysconfdir/logrotate.d/samba
@@ -910,8 +939,10 @@ TDB_NO_FSYNC=1 %make_build test
 %_libdir/samba/libutil-tdb-samba4.so
 
 %if_without libwbclient
-%_libdir/samba/libwbclient.so.*
+%ghost %_libdir/libwbclient.so.*
+%_libdir/samba/wbclient/libwbclient.so.*
 %_libdir/samba/libwinbind-client-samba4.so
+%_altdir/libwbclient-samba
 %endif # ! with_libwbclient
 
 %if_without libsmbclient
@@ -950,6 +981,7 @@ TDB_NO_FSYNC=1 %make_build test
 %attr(0700,root,root) %dir /var/log/samba/old
 %dir /var/run/samba
 %dir /var/run/winbindd
+%dir /var/lib/samba
 %attr(755,root,root) %dir %_localstatedir/cache/samba
 %attr(700,root,root) %dir /var/lib/samba/private
 %attr(755,root,root) %dir %_sysconfdir/samba
@@ -995,12 +1027,11 @@ TDB_NO_FSYNC=1 %make_build test
 %_sbindir/upgradeprovision
 %_libdir/mit_samba.so
 %_libdir/samba/bind9/dlz_bind9.so
-%_libdir/samba/libheimntlm-samba4.so.1
-%_libdir/samba/libheimntlm-samba4.so.1.0.1
-%_libdir/samba/libkdc-samba4.so.2
-%_libdir/samba/libkdc-samba4.so.2.0.0
-%_libdir/samba/libpac.so
-%_libdir/samba/gensec
+%_libdir/samba/libheimntlm-samba4.so.*
+%_libdir/samba/libkdc-samba4.so.*
+%_libdir/samba/libpac-samba4.so
+%dir %_libdir/samba/gensec
+%_libdir/samba/gensec/krb5.so
 %dir /var/lib/samba/sysvol
 %_datadir/samba/setup
 %_man8dir/samba.8*
@@ -1169,12 +1200,16 @@ TDB_NO_FSYNC=1 %make_build test
 
 %if_with libwbclient
 %files -n libwbclient
-%_libdir/libwbclient.so.*
+%ghost %_libdir/libwbclient.so.*
+%_libdir/samba/wbclient/libwbclient.so.*
 %_libdir/samba/libwinbind-client-samba4.so
+%_altdir/libwbclient-samba
 
 %files -n libwbclient-devel
 %_includedir/samba-4.0/wbclient.h
-%_libdir/libwbclient.so
+%ghost %_libdir/libwbclient.so
+%_libdir/samba/wbclient/libwbclient.so
+%_altdir/libwbclient-devel-samba
 %_pkgconfigdir/wbclient.pc
 %endif
 
@@ -1208,9 +1243,9 @@ TDB_NO_FSYNC=1 %make_build test
 
 %if_with testsuite
 # files to ignore in testsuite mode
-%_libdir/samba/libnss_wrapper.so
-%_libdir/samba/libsocket_wrapper.so
-%_libdir/samba/libuid_wrapper.so
+%_libdir/samba/libnss-wrapper.so
+%_libdir/samba/libsocket-wrapper.so
+%_libdir/samba/libuid-wrapper.so
 %endif
 
 %files test-libs
@@ -1321,6 +1356,10 @@ TDB_NO_FSYNC=1 %make_build test
 %endif
 
 %changelog
+* Wed Jul 15 2015 Alexey Shabalin <shaba@altlinux.ru> 4.2.3-alt1
+- 4.2.3
+- add alternatives for libwbclient
+
 * Mon Mar 23 2015 Alexey Shabalin <shaba@altlinux.ru> 4.2.0-alt1
 - 4.2.0
